requirements.txt
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
mysql-connector-python>=8.0.0
